- ok, this should enable Bobs portaudio fixes and ALSA depending on   --enable options.
This will have to be cleaned up a lot more to minimise #ifdef's
a2d_() is needed for portaudio which is needed by both windows and *nix
but only a small piece of start_threads.c is needed for portaudio
the ALSA part is #ifdef'd out when USE_PORTAUDIO is enabled.
audio_init.f90 and a2d.f90 now have #ifdef's on USE_PORTAUDIO
instead of Win32.

Lots to do still.



git-svn-id: svn://svn.berlios.de/wsjt/trunk@73 b3ca0830-8508-0410-8253-dac51d8a7955