- See http://gcc.gnu.org/onlinedocs/gcc-4.2.4/gfortran/BOZ-literal-constants.html#BOZ-literal-constants
Similarly, initializing an INTEGER variable with a statement such as DATA
i/Z'FFFFFFFF'/ will produce an integer overflow rather than the desired result
of -1 when i is a 32-bit integer on a system that supports 64-bit integers.
The `-fno-range-check' option can be used as a workaround for legacy code that
initializes integers in this manner. 

- This will have to be checked for g95, I might have to modify
  the FFLAGS depending on the compiler.



git-svn-id: svn://svn.berlios.de/wsjt/trunk@813 b3ca0830-8508-0410-8253-dac51d8a7955